src/com/wbiag/app/bo/ejb/actions/TeamTransferProcessActionTest.java
/*
 * Created on Jun 2, 2005
 *
 * TODO To change the template for this generated file go to
 * Window - Preferences - Java - Code Style - Code Templates
 */
package com.wbiag.app.bo.ejb.actions;

import java.util.Date;
import java.util.HashMap;
import java.util.Hashtable;
import java.util.Map;

import org.apache.log4j.Logger;

import junit.framework.TestSuite;

import com.wbiag.app.bo.InteractionTestCase;
import com.workbrain.app.bo.BOFormInstance;
import com.workbrain.app.workflow.AbstractActionProcess;
import com.workbrain.app.workflow.Action;
import com.workbrain.app.workflow.ActionResponse;
import com.workbrain.app.workflow.Branch;

/**
 * @author bhacko
 *
 * TODO To change the template for this generated type comment go to
 * Window - Preferences - Java - Code Style - Code Templates
 */
public class TeamTransferProcessActionTest extends InteractionTestCase {

    private static final Logger logger = Logger.getLogger(TeamTransferProcessActionTest.class);
    private AbstractActionProcess node = null;

    private static final String TEST_USER1 = "WORKBRAIN";
    private BOFormInstance form = null;
    private Action action = null;
    
    private static final String vId = "valueId";
    private static final String val = "value";
	
    public TeamTransferProcessActionTest(String testName) throws Exception {
        super(testName);
    }

    
    public void setUp() throws Exception {
        super.setUp();
        node = new TeamTransferProcessAction();
        // *** load workflow and form from DB
        form = createInteraction(
            TEST_USER1, "Team Transfer Workflow", "Team Transfer Request Common",
            false, false, new Date(), null);
        // *** set node parameters
        Map params = new Hashtable();
        action = super.getAction(params);
        action = new Action();

    }
	
    public void tearDown() throws Exception {
        super.tearDown();
        node = null;
        form = null;
        action = null;
    }
    
    public static TestSuite suite() {
        TestSuite result = new TestSuite();
        result.addTestSuite(TeamTransferProcessActionTest.class);
        return result;
    }
    
    public static void main(String[] args) {
        junit.textui.TestRunner.run(suite());
    }
    
    public void testProcessObjectWithRole() throws Exception {
    	runTestForm(
    			"MANAGER",
    			"3",
    			"NAC_45800F7,NAC_45800F6",
    			"10010,10012",
    			"IAG_TT001",
    			"10062",
    			"",
    			"",
    			"NAC_45800F3",
    			"10007",
    			"20050618 000000",
    			"20050615 000000",
    			"true",
    			"false",
    			"3",
    			"Internal"
    		);
    }

    public void testProcessObjectWithoutRole() throws Exception {
    	runTestForm(
        		"",
    			"",
    			"",
    			"",
    			"IAG_TT001",
    			"10062",
    			"",
    			"",
    			"NAC_45800F3",
    			"10007",
    			"20050618 000000",
    			"20050615 000000",
				"false",
    			"false",
    			"3",
    			"External"
    	);
    }
    
    private void runTestForm(
    		String strRoleName,
			String strRoleID,
			String strRoleTeamNames,
			String strRoleTeamIDs,
			String strEmpName,
			String strEmpID,
			String strStaffGroupName,
			String strStaffGroupID,
			String strDestTeamName,
			String strDestTeamID,
			String strDestEffDate,
			String strRoleEffDate,
			String strIsCurSuperTeam,
			String strIsLFSOEnabled,
			String strSuperRole,
			String strExpectedResult
			
		) throws Exception {
        ActionResponse result = null;

        Map roleValueMap = new HashMap();
        Map roleTeamsValueMap = new HashMap();
        Map empValueMap = new HashMap();
        Map staffGrpValueMap = new HashMap();
        Map destTeamValueMap = new HashMap();
        
        roleValueMap.put(val,strRoleName);
        roleValueMap.put(vId,strRoleID);
        
        roleTeamsValueMap.put(val, strRoleTeamNames);
        roleTeamsValueMap.put(vId,strRoleTeamIDs);

        empValueMap.put(val,strEmpName);
        empValueMap.put(vId,strEmpID);

        staffGrpValueMap.put(val,strStaffGroupName);
        staffGrpValueMap.put(vId,strStaffGroupID);
        
        destTeamValueMap.put(val,strDestTeamName);
        destTeamValueMap.put(vId,strDestTeamID);

        form.setFieldValue(TeamTransferProcessAction.PARAM_ROLE, roleValueMap);
        form.setFieldValue(TeamTransferProcessAction.PARAM_ROLE_TEAMS, roleTeamsValueMap);
        form.setFieldValue(TeamTransferProcessAction.PARAM_EMP, empValueMap);
        form.setFieldValue(TeamTransferProcessAction.PARAM_STAFF_GRP, staffGrpValueMap);
        form.setFieldValue(TeamTransferProcessAction.PARAM_DEST_TEAM, destTeamValueMap);
        
        form.setValue(TeamTransferProcessAction.PARAM_ROLE_EFF_DATE, strRoleEffDate);
        form.setValue(TeamTransferProcessAction.PARAM_DEST_EFF_DATE, strDestEffDate);
        form.setValue(TeamTransferProcessAction.LABEL_ROLE, "Assign Role: ");
        form.setValue(TeamTransferProcessAction.LABEL_ROLE_TEAMS, "Assign Role to Teams: ");
        form.setValue(TeamTransferProcessAction.LABEL_EMP, "Employee to Transfer: ");
        form.setValue(TeamTransferProcessAction.LABEL_STAFF_GRP, "Staff Group: ");
        form.setValue(TeamTransferProcessAction.LABEL_DEST_TEAM, "Transfer to Team: ");
        form.setValue(TeamTransferProcessAction.LABEL_ROLE_EFF_DATE, "Role Effective Date: ");
        form.setValue(TeamTransferProcessAction.LABEL_DEST_EFF_DATE, "Team Effective Date: ");

        form.setValue(TeamTransferProcessAction.PARAM_IS_CUR_SUPER_TEAM, strIsCurSuperTeam);
        form.setValue(TeamTransferProcessAction.PARAM_IS_LFSO_ENABLED, strIsLFSOEnabled);
        form.setValue(TeamTransferProcessAction.PARAM_SUPER_ROLE, strSuperRole);

//      *** set output branch names
        Branch[] branches = super.getBranches(new String[] {"Internal",
        		"External", "Failure"});
        // *** run test

        try {
        	result = node.processObject(action, form, branches, null);
        } catch (Exception e) {
        	e.printStackTrace();
        	logger.error(e);
        }
        //logger.debug(((Branch)result.getValue()).getLabel());
        assertTrue(true);
        //assertNotNull(result);
        //assertEquals(ActionResponse.BRANCH, result.getType());
        //assertEquals(strExpectedResult, ( (Branch) result.getValue()).getLabel());
    }
    
    
}
